Get position data from a specific tag file

Get position data from a specific tag file

Usage

read_pos_data(tag_id, file_connection, sep = ",", dec = ".")

Arguments

tag_id

The ID of the tag to retrieve position data for.

file_connection

FILE CONNECTION

sep

The separator used in the data file (default is comma).

dec

The decimal point character used in the data file (default is period).

Value

A data frame containing the position data with a POSIXct date_time column and position measurements.

Examples

get_pos_data("61029", "path/to/directory")
#> Error in get_pos_data("61029", "path/to/directory"): could not find function "get_pos_data"